+++ test/runtests.jl
@@ -24,7 +24,9 @@ using Test
     include("finite_field_braid.jl")
     include("fr_modp_semion.jl")
     include("fr_modp_fibonacci.jl")
+    include("fr_modp_ising.jl")
     include("braid_modp_fibonacci.jl")
+    include("braid_modp_ising.jl")
     include("zariski_diagnostics.jl")
     include("test_api_stability_docs.jl")
 end